% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapit.R
\name{mapit}
\alias{mapit}
\title{Quick Heat Map in ggplot2}
\usage{
mapit(dat, g = FALSE)
}
\arguments{
\item{dat}{The data frame you are using.

* coordinate columns must be named "x", "y";

* box width/height columns must be named "width", "height";

* statistic of interest must be names "statistic".}

\item{g}{TRUE or FALSE - Whether or not to include a legend, defaults to "FALSE"}
}
\value{
A ggplot2 heat map.
}
\description{
This function gives you a quick, basic heat map to use for data structured in the way described in the vignette.
}
\examples{
hitterVR <- varyres(hitter, mean, cutoff = 200, max = 6)
mapit(hitterVR[[4]])
}
