% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_res.R
\name{varyres}
\alias{varyres}
\title{A variable-resolution heat map generator}
\usage{
varyres(cutoff)
}
\arguments{
\item{cutoff}{Box subdivisions cease when a box sample size drops below the cutoff}
}
\description{
This function creates variable resolution heat maps accroding to a stopping rule
}
\examples{
data(hitter)
data <- varyres(hitter)
mapit(data[[4]])
}
